requirements.txt
Pillow==8.2.0
PySimpleGUI==4.38.0
